Fail fast for integration tests
- Integration tesing should stop whenever there is a failure
- Add ability to create iml diagnostics with a provisioning script
- Add pub key to authorized keys for diagnostics upload
- Create diagnostic scripts in parallel on failure
