+++ iml-system-test-utils/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["IML Team <iml@whamcloud.com>"]
 edition = "2018"
 
 [dependencies]
+async-trait = "0.1"
 futures = "0.3"
 iml-cmd = { version = "0.2.0", path = "../iml-cmd" }
 iml-fs = { version = "0.2", path = "../iml-fs" }
