+++ vagrant/Vagrantfile
@@ -77,7 +77,7 @@ Vagrant.configure('2') do |config|
 
   config.vm.provision 'shell', path: './scripts/disable_selinux.sh'
 
-  system("ssh-keygen -t rsa -m PEM -N '' -f id_rsa") unless File.exist?('id_rsa')
+  system("ssh-keygen -t rsa -m PEM -N '' -f id_rsa && cat id_rsa.pub >> ~/.ssh/authorized_keys") unless File.exist?('id_rsa')
 
   config.vm.provision 'ssh', type: 'shell', path: './scripts/key_config.sh'
 
@@ -128,6 +128,8 @@ Vagrant.configure('2') do |config|
     provision_mgmt_net adm, '10'
 
     provision_fence_agents adm
+
+    create_iml_diagnostics adm
 
       adm.vm.synced_folder '../',
                            '/integrated-manager-for-lustre/',
@@ -222,6 +224,8 @@ Vagrant.configure('2') do |config|
       mds.vm.provider 'virtualbox' do |vbx|
         vbx.name = "mds#{i}"
       end
+
+      create_iml_diagnostics mds
 
       provision_lnet_net mds, "1#{i}"
 
@@ -449,6 +453,8 @@ Vagrant.configure('2') do |config|
       oss.vm.provider 'virtualbox' do |vbx|
         vbx.name = "oss#{i}"
       end
+
+      create_iml_diagnostics oss
 
       # Lustre / application network
       provision_lnet_net oss, "2#{i}"
@@ -1071,5 +1077,13 @@ def wait_for_ntp_docker(config)
                           type: 'shell',
                           run: 'never',
                           path: 'scripts/wait_for_ntp.sh',
+                          args: ["10.73.10.1"]
+end
+
+def create_iml_diagnostics(config)
+  config.vm.provision 'create-iml-diagnostics',
+                          type: 'shell',
+                          run: 'never',
+                          path: 'scripts/create_iml_diagnostics.sh',
                           args: ["10.73.10.1"]
 end
